move is_agent_alive to amqp_client
This function is rarely used, so let's not put it in the main utils
module - it requires importing amqp_client anyway, which pulls in
pika, which is not small.

Now, utils doesn't need pika, nor does it need cloudify.amqp_client
