+++ cloudify/constants.py
@@ -117,3 +117,4 @@ TERMINATED_STATES = [
 ]
 
 CONVENTION_APPLICATION_BLUEPRINT_FILE = 'blueprint.yaml'
+INSPECT_TIMEOUT = 30
